requirements.txt
numpy
keras
tensorflow
tensorflow-gpu
opencv-python
beautifulsoup4==4.7.1
matplotlib
sklearn
pillow
tqdm
lxml
pascal_voc_writer